requirements.txt
Django==1.11.1
django-cors-headers==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
pipenv==10.1.2
postgres==2.2.1
psycopg2==2.7.1
whitenoise==3.3.1
coverage==4.5.1
dj-database-url==0.4.2
django-star-ratings